refactor: extract the path profile country and heatscale func into a custom hooks
